// apps/frontend/mobile/src/api/generated/mobile-gateway/models/formatEnum.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Blockout Mobile Gateway API
 * OpenAPI spec version: 2.0.0
 */

/**
 * Supported Blockout team formats.
 */
export type FormatEnum = (typeof FormatEnum)[keyof typeof FormatEnum];

export const FormatEnum = {
  SIX: 'SIX',
  FOUR: 'FOUR',
  TWO: 'TWO',
} as const;